+++ deployments/chef/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 🚩 Deprecations 🚩
+
+- Fluentd support has been deprecated and will be removed in a future release.
+  Please refer to [deprecation documentation](../../docs/deprecations/fluentd-support.md) for more information ([#6339](https://github.com/signalfx/splunk-otel-collector/pull/6339))
+
 ## chef-v0.17.0
 
 - Bug fix: Fix bug that caused custom variables to not be set when running on Windows